Add files via upload
We have a geographical locations , fare amount and time. It is a simple regression problem. Using data preprocessing techniques data is cleaned, visualization is done to visualize any interesting patterns in the data. Finally a model is selected which can give the highest accuracy without overfitting or underfitting of data.